Moved search components our of src/ folder.
