+++ app/javascript/Search/index.js
