+++ app/javascript/packs/Search.jsx
@@ -1,6 +1,6 @@
 import { h, render } from 'preact';
-import { Search } from '../src/components/Search';
-import 'focus-visible'
+import { Search } from '../Search';
+import 'focus-visible';
 
 document.addEventListener('DOMContentLoaded', () => {
   const root = document.getElementById('top-bar--search');
